requirements.txt
Flask==3.0.3
nltk==3.8.1
scikit-learn==1.4.2
google-generativeai==0.8.5
python-dotenv==1.0.1
